+++ src/main/java/org/kdb/inside/brains/QKeyword.java
@@ -1,13 +1,19 @@
 package org.kdb.inside.brains;
 
 public final class QKeyword {
-    final String name;
-    final String description;
     final Type type;
+    final String name;
+    final String arguments;
+    final String description;
 
     QKeyword(String name, Type type, String description) {
+        this(name, type, null, description);
+    }
+
+    QKeyword(String name, Type type, String arguments, String description) {
         this.name = name;
         this.type = type;
+        this.arguments = arguments;
         this.description = description;
     }
 
@@ -17,6 +23,10 @@ public final class QKeyword {
 
     public String getName() {
         return name;
+    }
+
+    public String getArguments() {
+        return arguments;
     }
 
     public String getDescription() {